Dumpvalue: Generalize for EBCDIC
I overlooked this module until now.  It turns out that much of the code
I had changed had a common ancestor with the code I had already changed
to work on EBCDIC in lib/dumpvar.pl.  So I just copied that changing the
things that needed to be different.  It appears that Dumpvalue had a bug
in that it did not escape NUL, of all the C0 controls.  I changed it to
do so.