XXX Extend UTF-EBCDIC to handle up to 2**64-1
This uses the same mechanism that Perl does to extend UTF-8 beyond what
might be its natural maximum.  That is, when the UTF-8 start byte is
0xFF, it adds 6 more bytes to the character.  For this purpose, only 1
byte was needed, but using 6 makes the handling between the two
character sets uniform, so some macros can now be shared instead of
having different values, and I thought there might be some subtetly that
I am unaware of, but which tripped up earlier implementors.  I did
notice that if I added fewer bytes, there was a lot more possibility of
overlongs.  XXX Example from utf8.out